+++ test/src/src-test.js
@@ -1,4 +1,4 @@
 const runTests = require('../tests')
-const { validate, validateFunc, type } = require('../../src')
+const { validate, validateFunc, type, typeClass } = require('../../src')
 
-runTests(validate, validateFunc, type)
+runTests(validate, validateFunc, type, typeClass)