+++ docs/changelog.md
@@ -1,5 +1,12 @@
 # Change Log
 
+### V 2.0.3
+GH-7 Add instance and strict-instance validation types
+
+GH-3 Add "type" api option
+
+GH-2 API to use JS primatives instead of strings
+
 ### V 2.0.2
 Non-code admin changes. Add contribution guidelines, change log, and branding. Update keywords.